Corrected RPM from uint8 to int32

+++ sw/airborne/modules/meteo/ekf_aw_wrapper.h
@@ -21,8 +21,8 @@ struct ekfAw {
   struct FloatRates gyro;   ///< Last gyroscope measurements
   struct FloatEulers euler; /// Euler angles
   
-  uint16_t last_RPM_hover[4]; // Value obtained from ABI Callback
-  uint16_t last_RPM_pusher; // Value obtained from ABI Callback
+  int32_t last_RPM_hover[4]; // Value obtained from ABI Callback
+  int32_t last_RPM_pusher; // Value obtained from ABI Callback
   float RPM_hover[4]; /// Hover motor RPM
   float RPM_pusher; /// Pusher motor RPM
   float skew; /// Skew
